// src/types/Responses.ts
// No AI was used in the creation of this project.
// Note: The name "EightBall.tsx," among others, reflect an earlier state of this
//       project. The current project has nothing to do with Magic 8-Balls.

export type BallResponse={
    id: number, // sort of new (retrieved but unused in example)
    name: string,
    status: string,
    species: string,
    type: string, // new
    gender: string, // new
    image: string,
    created: string // new
}